@@ -53,7 +53,12 @@ func init() {
 	}
 }
 
-func (p *Plugin) getCpuLoad(params []string) (result interface{}, err error) {
+// Period returns 1, for a required interface call for interface Collector.
+func (*Plugin) Period() int {
+	return 1
+}
+
+func (*Plugin) getCPULoad(_ []string) (any, error) {
 	return nil, plugin.UnsupportedMetricError
 }
 
@@ -138,6 +143,10 @@ func (p *Plugin) addCpu(index int) {
 			p.cpus = append(p.cpus, &cpuUnit{index: idx + 1, status: cpuStatusOffline})
 		}
 	}
+}
+
+func (*Plugin) getCounterAverage(cpu *cpuUnit, counter cpuCounter, period historyIndex) any {
+	return cpu.counterAverage(counter, period, 1)
 }
 
 func numCPUConf() int {
