@@ -16,6 +16,8 @@ package cpu
 
 import (
 	"errors"
+	"sync"
+	"time"
 	"unsafe"
 
 	"golang.zabbix.com/agent2/pkg/win32"
@@ -37,6 +39,8 @@ type Plugin struct {
 	plugin.Base
 	cpus      []*cpuUnit
 	collector *pdhCollector
+	cpusMu    sync.Mutex
+	stop      chan struct{}
 }
 
 func init() {
@@ -89,7 +93,7 @@ func numCPU() (numCpu int) {
 	return
 }
 
-func (p *Plugin) getCpuLoad(params []string) (result interface{}, err error) {
+func (p *Plugin) getCPULoad(params []string) (result any, err error) {
 	split := 1
 
 	period := historyIndex(defaultIndex)
@@ -113,10 +117,13 @@ func (p *Plugin) getCpuLoad(params []string) (result interface{}, err error) {
 		return nil, zbxerr.ErrorTooManyParameters
 	}
 
+	p.cpusMu.Lock()
+	defer p.cpusMu.Unlock()
+
 	return p.cpus[0].counterAverage(counterLoad, period, split), nil
 }
 
-func (p *Plugin) Collect() (err error) {
+func (p *Plugin) collectCpuData() (err error) {
 	ok, err := p.collector.collect()
 	if err != nil || !ok {
 		return
@@ -131,6 +138,7 @@ func (p *Plugin) Collect() (err error) {
 		}
 		slot.util += p.collector.cpuUtil(i)
 
+		p.cpusMu.Lock()
 		if cpu.tail = cpu.tail.inc(); cpu.tail == cpu.head {
 			cpu.head = cpu.head.inc()
 		}
@@ -138,6 +146,7 @@ func (p *Plugin) Collect() (err error) {
 		// can be added to it resulting in incrementing counter
 		nextSlot := &cpu.history[cpu.tail]
 		*nextSlot = *slot
+		p.cpusMu.Unlock()
 	}
 	return
 }
@@ -150,11 +159,36 @@ func (p *Plugin) Start() {
 	}
 	p.cpus = p.newCpus(numCpus)
 	p.collector.open(numCpus, numGroups)
+
+	p.stop = make(chan struct{})
+
+	go func() {
+		t := time.NewTicker(1 * time.Second)
+		defer t.Stop()
+		for {
+			select {
+			case <-p.stop:
+				return
+			case <-t.C:
+				p.Debugf("starting to collect CPU performance data")
+
+				err := p.collectCpuData()
+				if err != nil {
+					p.Warningf("failed to get CPU performance data: '%s'", err)
+					continue
+				}
+
+				p.Debugf("collected CPU performance data")
+			}
+		}
+	}()
 }
 
 func (p *Plugin) Stop() {
 	p.collector.close()
 	p.cpus = nil
+
+	p.stop <- struct{}{}
 }
 
 func (p *Plugin) Export(key string, params []string, ctx plugin.ContextProvider) (result interface{}, err error) {
@@ -166,7 +200,7 @@ func (p *Plugin) Export(key string, params []string, ctx plugin.ContextProvider)
 	case "system.cpu.discovery":
 		return p.getCpuDiscovery(params)
 	case "system.cpu.load":
-		return p.getCpuLoad(params)
+		return p.getCPULoad(params)
 	case "system.cpu.num":
 		if len(params) > 0 && params[0] == "max" {
 			return nil, errors.New("Invalid first parameter.")
@@ -177,4 +211,10 @@ func (p *Plugin) Export(key string, params []string, ctx plugin.ContextProvider)
 	default:
 		return nil, plugin.UnsupportedMetricError
 	}
+}
+
+func (p *Plugin) getCounterAverage(cpu *cpuUnit, counter cpuCounter, period historyIndex) any {
+	p.cpusMu.Lock()
+	defer p.cpusMu.Unlock()
+	return cpu.counterAverage(counter, period, 1)
 }
