@@ -20,6 +20,7 @@ package win32
 import (
 	"errors"
 	"fmt"
+	"sync"
 	"syscall"
 	"time"
 	"unsafe"
@@ -64,6 +65,10 @@ var (
 	pdhRemoveCounter            = hPdh.mustGetProcAddress("PdhRemoveCounter")
 	pdhEnumObjectItems          = hPdh.mustGetProcAddress("PdhEnumObjectItemsW")
 	pdhEnumObjects              = hPdh.mustGetProcAddress("PdhEnumObjectsW")
+
+	// mutex to prevent concurrent calls of Windows API PDH functions when one of the following is already being executed
+	// pdhEnumObjectItems(), pdhEnumObjects() and pdhCollectQueryData()
+	pdhMu sync.RWMutex
 )
 
 type Instance struct {
@@ -76,6 +81,8 @@ func PdhOpenQuery(dataSource *string, userData uintptr) (query PDH_HQUERY, err e
 		wcharSource := windows.StringToUTF16(*dataSource)
 		source = uintptr(unsafe.Pointer(&wcharSource[0]))
 	}
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 	ret, _, _ := syscall.Syscall(pdhOpenQuery, 3, source, userData, uintptr(unsafe.Pointer(&query)))
 
 	if syscall.Errno(ret) != windows.ERROR_SUCCESS {
@@ -85,6 +92,8 @@ func PdhOpenQuery(dataSource *string, userData uintptr) (query PDH_HQUERY, err e
 }
 
 func PdhCloseQuery(query PDH_HQUERY) (err error) {
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 	ret, _, _ := syscall.Syscall(pdhCloseQuery, 1, uintptr(query), 0, 0)
 	if syscall.Errno(ret) != windows.ERROR_SUCCESS {
 		return newPdhError(ret)
@@ -94,6 +103,8 @@ func PdhCloseQuery(query PDH_HQUERY) (err error) {
 
 func PdhAddCounter(query PDH_HQUERY, path string, userData uintptr) (counter PDH_HCOUNTER, err error) {
 	wcharPath, _ := syscall.UTF16PtrFromString(path)
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 	ret, _, _ := syscall.Syscall6(pdhAddCounter, 4, uintptr(query), uintptr(unsafe.Pointer(wcharPath)), userData,
 		uintptr(unsafe.Pointer(&counter)), 0, 0)
 
@@ -105,6 +116,8 @@ func PdhAddCounter(query PDH_HQUERY, path string, userData uintptr) (counter PDH
 
 func PdhAddEnglishCounter(query PDH_HQUERY, path string, userData uintptr) (counter PDH_HCOUNTER, err error) {
 	wcharPath, _ := syscall.UTF16PtrFromString(path)
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 	ret, _, _ := syscall.Syscall6(pdhAddEnglishCounter, 4, uintptr(query), uintptr(unsafe.Pointer(wcharPath)), userData,
 		uintptr(unsafe.Pointer(&counter)), 0, 0)
 
@@ -115,6 +128,8 @@ func PdhAddEnglishCounter(query PDH_HQUERY, path string, userData uintptr) (coun
 }
 
 func PdhCollectQueryData(query PDH_HQUERY) (err error) {
+	pdhMu.Lock()
+	defer pdhMu.Unlock()
 	ret, _, _ := syscall.Syscall(pdhCollectQueryData, 1, uintptr(query), 0, 0)
 	if syscall.Errno(ret) != windows.ERROR_SUCCESS {
 		return newPdhError(ret)
@@ -144,19 +159,21 @@ func PdhGetFormattedCounterValueDouble(counter PDH_HCOUNTER, tryCount int) (*flo
 }
 
 func PdhGetFormattedCounterValueInt64(counter PDH_HCOUNTER) (value *int64, err error) {
-	return PdhGetFormattedCounterValueInt64Helper(counter, true)
+	return pdhGetFormattedCounterValueInt64Helper(counter, true)
 }
 
-func PdhGetFormattedCounterValueInt64Helper(counter PDH_HCOUNTER, retry bool) (value *int64, err error) {
+func pdhGetFormattedCounterValueInt64Helper(counter PDH_HCOUNTER, retry bool) (value *int64, err error) {
 	var pdhValue PDH_FMT_COUNTERVALUE_LARGE
+	pdhMu.RLock()
 	ret, _, _ := syscall.Syscall6(pdhGetFormattedCounterValue, 4, uintptr(counter), uintptr(PDH_FMT_LARGE), 0,
 		uintptr(unsafe.Pointer(&pdhValue)), 0, 0)
+	pdhMu.RUnlock()
 	if syscall.Errno(ret) != windows.ERROR_SUCCESS {
 		if ret == PDH_CALC_NEGATIVE_DENOMINATOR {
 			if retry {
 				log.Debugf("Detected performance counter with negative denominator, retrying in 1 second")
 				time.Sleep(time.Second)
-				return PdhGetFormattedCounterValueInt64Helper(counter, false)
+				return pdhGetFormattedCounterValueInt64Helper(counter, false)
 			}
 			log.Warningf("Detected performance counter with negative denominator the second time after retry, giving up...")
 		} else if ret == PDH_INVALID_DATA || ret == PDH_CSTATUS_INVALID_DATA {
@@ -168,6 +185,8 @@ func PdhGetFormattedCounterValueInt64Helper(counter PDH_HCOUNTER, retry bool) (v
 }
 
 func PdhRemoveCounter(counter PDH_HCOUNTER) (err error) {
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 	ret, _, _ := syscall.Syscall(pdhRemoveCounter, 1, uintptr(counter), 0, 0)
 	if syscall.Errno(ret) != windows.ERROR_SUCCESS {
 		return newPdhError(ret)
@@ -179,6 +198,8 @@ func PdhParseCounterPath(path string) (elements *PDH_COUNTER_PATH_ELEMENTS, err 
 	wPath := windows.StringToUTF16(path)
 	ptrPath := uintptr(unsafe.Pointer(&wPath[0]))
 	var size uint32
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 	ret, _, _ := syscall.Syscall6(pdhParseCounterPath, 4, ptrPath, 0, uintptr(unsafe.Pointer(&size)), 0, 0, 0)
 	if ret != PDH_MORE_DATA && syscall.Errno(ret) != windows.ERROR_SUCCESS {
 		return nil, newPdhError(ret)
@@ -194,6 +215,8 @@ func PdhParseCounterPath(path string) (elements *PDH_COUNTER_PATH_ELEMENTS, err 
 }
 
 func PdhMakeCounterPath(elements *PDH_COUNTER_PATH_ELEMENTS) (path string, err error) {
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 	size := uint32(PDH_MAX_COUNTER_NAME)
 	buf := make([]uint16, size)
 	ret, _, _ := syscall.Syscall6(pdhMakeCounterPath, 4, uintptr(unsafe.Pointer(elements)),
@@ -205,6 +228,8 @@ func PdhMakeCounterPath(elements *PDH_COUNTER_PATH_ELEMENTS) (path string, err e
 }
 
 func PdhLookupPerfNameByIndex(index int) (path string, err error) {
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 	size := uint32(PDH_MAX_COUNTER_NAME)
 	buf := make([]uint16, size)
 	ret, _, _ := syscall.Syscall6(pdhLookupPerfNameByIndex, 4, 0, uintptr(index), uintptr(unsafe.Pointer(&buf[0])),
@@ -221,6 +246,8 @@ func PdhLookupPerfIndexByName(name string) (idx int, err error) {
 		return 0, err
 	}
 
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 	ret, _, _ := syscall.Syscall(pdhLookupPerfIndexByName, 3, 0, uintptr(unsafe.Pointer(nameUTF16)), uintptr(unsafe.Pointer(&idx)))
 	if syscall.Errno(ret) != windows.ERROR_SUCCESS {
 		return 0, newPdhError(ret)
@@ -230,6 +257,8 @@ func PdhLookupPerfIndexByName(name string) (idx int, err error) {
 }
 
 func PdhEnumObjectItems(objectName string) (instances []Instance, err error) {
+	pdhMu.Lock()
+	defer pdhMu.Unlock()
 	var counterListSize, instanceListSize uint32
 	nameUTF16, err := syscall.UTF16FromString(objectName)
 	if err != nil {
@@ -295,6 +324,8 @@ func PdhEnumObjectItems(objectName string) (instances []Instance, err error) {
 }
 
 func PdhEnumObject() (objects []string, err error) {
+	pdhMu.Lock()
+	defer pdhMu.Unlock()
 	var objectListSize uint32
 	ret, _, _ := syscall.Syscall6(pdhEnumObjects, 6, 0, 0, 0, uintptr(unsafe.Pointer(&objectListSize)),
 		uintptr(PERF_DETAIL_WIZARD), bool2uintptr(true))
@@ -329,6 +360,9 @@ func PdhEnumObject() (objects []string, err error) {
 
 func getCounterValueDouble(counter PDH_HCOUNTER) (*float64, error) {
 	var pdhValue PDH_FMT_COUNTERVALUE_DOUBLE
+
+	pdhMu.RLock()
+	defer pdhMu.RUnlock()
 
 	ret, _, _ := syscall.SyscallN(
 		pdhGetFormattedCounterValue,
