@@ -26,6 +26,7 @@
 #include "zbxnum.h"
 #include "zbxstr.h"
 #include "zbxalgo.h"
+#include "zbxcachehistory.h"
 
 #define ZBX_VMWARE_DATASTORE_SIZE_TOTAL		0
 #define ZBX_VMWARE_DATASTORE_SIZE_FREE		1
@@ -1063,6 +1064,9 @@ static void	vmware_get_events(const zbx_vector_vmware_event_ptr_t *events,
 		const zbx_vmware_eventlog_state_t *evt_state, const zbx_dc_item_t *item,
 		zbx_vector_agent_result_ptr_t *add_results)
 {
+	zbx_uint64_t	key;
+	time_t		timestamp;
+
 	zabbix_log(LOG_LEVEL_DEBUG, "In %s() last_key:" ZBX_FS_UI64 " last_ts:" ZBX_FS_TIME_T " events:%d top event id:"
 			ZBX_FS_UI64 " top event ts:" ZBX_FS_TIME_T, __func__, evt_state->last_key, evt_state->last_ts,
 			events->values_num, events->values[0]->key, events->values[0]->timestamp);
@@ -1082,7 +1086,8 @@ static void	vmware_get_events(const zbx_vector_vmware_event_ptr_t *events,
 
 		if (SUCCEED == zbx_set_agent_result_type(add_result, item->value_type, event->message))
 		{
-			zbx_set_agent_result_meta(add_result, event->key, 0);
+			key = event->key;
+			timestamp = event->timestamp;
 
 			if (ITEM_VALUE_TYPE_LOG == item->value_type)
 			{
@@ -1095,6 +1100,9 @@ static void	vmware_get_events(const zbx_vector_vmware_event_ptr_t *events,
 		else
 			zbx_free(add_result);
 	}
+
+	if (0 != add_results->values_num)
+		zbx_set_agent_result_meta(add_results->values[add_results->values_num - 1], key, (int)timestamp);
 
 	zabbix_log(LOG_LEVEL_DEBUG, "End of %s(): events:%d", __func__, add_results->values_num);
 }
@@ -1173,9 +1181,9 @@ int	check_vcenter_eventlog(AGENT_REQUEST *request, const zbx_dc_item_t *item, AG
 	unsigned char		skip_old, severity = 0;
 	zbx_vmware_service_t	*service;
 	int			ret = SYSINFO_RET_FAIL;
-	time_t			lastaccess;
+	double			hc_pused = 0;
 
-	zabbix_log(LOG_LEVEL_DEBUG, "In %s()", __func__);
+	zabbix_log(LOG_LEVEL_DEBUG, "In %s() lastlogsize:" ZBX_FS_UI64, __func__, request->lastlogsize);
 
 	if (3 < request->nparam || 0 == request->nparam)
 	{
@@ -1216,16 +1224,14 @@ int	check_vcenter_eventlog(AGENT_REQUEST *request, const zbx_dc_item_t *item, AG
 	if (NULL == (service = get_vmware_service(url, item->username, item->password, result, &ret)))
 		goto unlock;
 
-	lastaccess = time(NULL);
-
 	if (0 != service->eventlog.lastaccess &&
-			service->eventlog.interval != lastaccess - service->eventlog.lastaccess)
+			service->eventlog.interval != service->lastaccess - service->eventlog.lastaccess)
 	{
 		service->jobs_flag |= ZBX_VMWARE_REQ_UPDATE_EVENTLOG;
-		service->eventlog.interval = lastaccess - service->eventlog.lastaccess;
+		service->eventlog.interval = service->lastaccess - service->eventlog.lastaccess;
 	}
 
-	service->eventlog.lastaccess = lastaccess;
+	service->eventlog.lastaccess = service->lastaccess;
 
 	if (0 == (service->jobs_flag & ZBX_VMWARE_UPDATE_EVENTLOG))
 		service->jobs_flag |= ZBX_VMWARE_REQ_UPDATE_EVENTLOG;
@@ -1238,7 +1244,7 @@ int	check_vcenter_eventlog(AGENT_REQUEST *request, const zbx_dc_item_t *item, AG
 	{
 		/* this may happen if recreate item vmware.eventlog for same service URL */
 		service->eventlog.last_key = request->lastlogsize;
-		service->eventlog.last_ts = 0;
+		service->eventlog.last_ts = request->mtime;
 		service->eventlog.skip_old = skip_old;
 		service->eventlog.owner_itemid = item->itemid;
 	}
@@ -1262,6 +1268,11 @@ int	check_vcenter_eventlog(AGENT_REQUEST *request, const zbx_dc_item_t *item, AG
 		SET_MSG_RESULT(result, zbx_strdup(NULL, service->eventlog.data->error));
 		goto unlock;
 	}
+	else if (80 < (hc_pused = zbx_hc_mem_pused_lock()))
+	{
+		zabbix_log(LOG_LEVEL_DEBUG,"%s():eventlog data is suspended due to history cache is overloaded:%.2f%%",
+				__func__, hc_pused);
+	}
 	else if (0 < service->eventlog.data->events.values_num)
 	{
 		/* Some times request->lastlogsize value gets stuck due to concurrent update of history cache */
@@ -1276,7 +1287,9 @@ int	check_vcenter_eventlog(AGENT_REQUEST *request, const zbx_dc_item_t *item, AG
 unlock:
 	zbx_vmware_unlock();
 out:
-	zabbix_log(LOG_LEVEL_DEBUG, "End of %s():%s", __func__, zbx_sysinfo_ret_string(ret));
+	zabbix_log(LOG_LEVEL_DEBUG, "End of %s():%s events:%d hc_pused:%.2f%% last_key:" ZBX_FS_UI64, __func__,
+			zbx_sysinfo_ret_string(ret), add_results->values_num, hc_pused, 0 != add_results->values_num ?
+			add_results->values[add_results->values_num - 1]->lastlogsize : 0);
 
 	return ret;
 }
