@@ -135,6 +135,60 @@ int	vmware_shared_is_ready(void)
 	return SUCCEED;
 }
 
+/******************************************************************************
+ *                                                                            *
+ * Purpose: getting size of shared memory available for events of VC instance *
+ *                                                                            *
+ ******************************************************************************/
+float	vmware_shared_evtpart_size(const int num)
+{
+#	define	DEFAULT_FACTOR(n)	((float)1/n)
+	int	i, total = 0, vc_active = 0;
+
+	for (i = 0; i < vmware->services.values_num; i++)
+	{
+		if (0 != (vmware->services.values[i]->state & ZBX_VMWARE_STATE_FAILED))
+			continue;
+
+		vc_active++;
+	}
+
+	if (1 >= vc_active)
+		return 1;
+
+	if (0 == num)
+		return DEFAULT_FACTOR(vc_active);
+
+	for (i = 0; i < vmware->services.values_num; i++)
+	{
+		if (0 == (vmware->services.values[i]->state & ZBX_VMWARE_STATE_SHMEM_READY))
+			return DEFAULT_FACTOR(vc_active);
+
+		if (NULL == vmware->services.values[i]->eventlog.data)
+			return DEFAULT_FACTOR(vc_active);
+
+		total += vmware->services.values[i]->eventlog.expect_num;
+	}
+
+	if (0 != total)
+	{
+#		define	MIN_FACTOR	(((float)1 / vc_active) / 4)
+		float	factor = (float)num / total;
+
+		if (factor < MIN_FACTOR)
+			factor = MIN_FACTOR;
+		else if (factor > (1 - MIN_FACTOR * (vc_active - 1)))
+			factor = 1 - MIN_FACTOR * (vc_active - 1);
+
+		return factor;
+#		undef	MIN_FACTOR
+	}
+	else
+		return DEFAULT_FACTOR(vc_active);
+
+#	undef	DEFAULT_FACTOR
+}
+
 /* the vmware resource pool chunk */
 typedef struct
 {
@@ -2867,6 +2921,7 @@ zbx_vmware_service_t	*zbx_vmware_get_service(const char* url, const char* userna
 	service->eventlog.req_sz = 0;
 	service->eventlog.oom = 0;
 	service->eventlog.job_revision = 0;
+	service->eventlog.expect_num = 0;
 	service->jobs_num = 0;
 	vmware_shmem_vector_vmware_entity_tags_ptr_create_ext(&service->data_tags.entity_tags);
 	service->data_tags.error = NULL;
@@ -3088,7 +3143,7 @@ static void	zbx_vmware_job_create(zbx_vmware_t *vmw, zbx_vmware_service_t *servi
  ******************************************************************************/
 static void	zbx_vmware_jobs_create(zbx_vmware_t *vmw, zbx_vmware_service_t *service)
 {
-	int	req_flag = 0x1, jobs_req = service->jobs_flag >> ZBX_VMWARE_REQ;
+	int	req_flag = 0x1, jobs_req = ((service->jobs_flag & ZBX_VMWARE_REQ_MASK) >> ZBX_VMWARE_REQ);
 
 	while (0 != jobs_req)
 	{
