@@ -24,7 +24,6 @@
 #include "zbxcacheconfig.h"
 #include "zbxdb.h"
 #include "zbxhistory.h"
-#include "zbxshmem.h"
 #include "zbxtime.h"
 #include "zbxtrends.h"
 #include "zbx_item_constants.h"
@@ -1697,8 +1696,7 @@ int	zbx_hc_check_proxy(zbx_uint64_t proxyid)
 
 	zbx_dbcache_lock();
 
-	hc_pused = 100 * (double)(zbx_dbcache_get_hc_mem()->total_size - zbx_dbcache_get_hc_mem()->free_size) /
-			zbx_dbcache_get_hc_mem()->total_size;
+	hc_pused = zbx_hc_mem_pused();
 
 	if (20 >= hc_pused)
 	{
