@@ -47,15 +47,13 @@ class CWidgetFieldColumnsListView extends CWidgetFieldView {
 		];
 
 		$row_template = new CTemplateTag($this->field->getName().'-row-tmpl', new CRow([
-			(new CCol((new CDiv)->addClass(ZBX_STYLE_DRAG_ICON)))->addClass(ZBX_STYLE_TD_DRAG_ICON),
 			(new CDiv('#{items}'))->addClass('text'),
 			(new CList(array_merge($row_actions, [(new CSpan())->addClass('js-column-data')])))
 				->addClass(ZBX_STYLE_HOR_LIST)
 		]));
 
 		$header = [
-			$row_template,
-			(new CColHeader(_('Patterns')))->addStyle('width: 100%'),
+			(new CColHeader(_('Patterns')))->addStyle('width: 100%')->addItem($row_template),
 			_('Actions')
 		];
 
@@ -76,7 +74,6 @@ class CWidgetFieldColumnsListView extends CWidgetFieldView {
 			}
 
 			$view->addRow((new CRow([
-				(new CCol((new CDiv)->addClass(ZBX_STYLE_DRAG_ICON)))->addClass(ZBX_STYLE_TD_DRAG_ICON),
 				(new CDiv($items))->addClass('text'),
 				(new CList(array_merge($row_actions, [(new CSpan($column_data))->addClass('js-column-data')])))
 					->addClass(ZBX_STYLE_HOR_LIST)
